+++ funding/urls.py
@@ -5,8 +5,8 @@ from .views import (
     CampaignListView,
     CampaignDetailView,
     CreateDonationView,
-    CampaignCreateView,
-    OrganisationListView
+    OrganisationListView,
+    DonationDetailView
 )
 from .admin_views import (
     org_owner_test_view
@@ -17,7 +17,7 @@ urlpatterns = [
     path('campaigns/', CampaignListView.as_view(), name='campaign_list'),
     path('campaigns/<int:pk>/', CampaignDetailView.as_view(), name='campaign_detail'),
     path('campaigns/<int:pk>/donate/', CreateDonationView.as_view(), name='campaign_donate'),
-    path('campaigns/new/', CampaignCreateView.as_view(), name='campaign_new'),
+    path('donations/<int:pk>/', DonationDetailView.as_view(), name='donation_detail'),
 
     # Test URL for role_required decorator
     path('test/org-owner-only/', org_owner_test_view, name='test_org_owner_view'),